changed Webdev to WebDev, factored out alumni/team containers
